Final stage project structure refactoring
