+++ Wox.Plugin.SystemPlugins/FileSystem/FileSystemSettings.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="Wox.Plugin.SystemPlugins.FileSystem.FileSystemSettings"
+﻿<UserControl x:Class="Wox.Plugins.Internal.FileSystem.FileSystemSettings"
 			 xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 			 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
