+++ Wox.Plugin.SystemPlugins/Program/ProgramSourceSetting.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="Wox.Plugin.SystemPlugins.Program.ProgramSourceSetting"
+﻿<Window x:Class="Wox.Plugins.Internal.Program.ProgramSourceSetting"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         ResizeMode="NoResize"
