+++ Wox.Plugin.SystemPlugins/Program/ProgramSourceSetting.xaml.cs
@@ -4,9 +4,8 @@ using System.Linq;
 using System.Windows;
 using System.Windows.Controls;
 using Wox.Core.Data.UserSettings;
-using MessageBox = System.Windows.MessageBox;
 
-namespace Wox.Plugin.SystemPlugins.Program
+namespace Wox.Plugins.Internal.Program
 {
     public partial class ProgramSourceSetting : Window
     {
