+++ Wox.Plugin.SystemPlugins/Program/ProgramSources/CommonStartMenuProgramSource.cs
@@ -3,7 +3,7 @@ using System.Runtime.InteropServices;
 using System.Text;
 using Wox.Core.Data.UserSettings;
 
-namespace Wox.Plugin.SystemPlugins.Program.ProgramSources
+namespace Wox.Plugins.Internal.Program.ProgramSources
 {
     [global::System.ComponentModel.Browsable(false)]
     public class CommonStartMenuProgramSource : FileSystemProgramSource
