+++ Wox.Plugin.SystemPlugins/Program/Programs.cs
@@ -3,10 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using Wox.Core;
 using Wox.Core.Data.UserSettings;
-using Wox.Plugin.SystemPlugins.Program.ProgramSources;
-using Wox.Plugins;
+using Wox.Plugins.Internal.Program.ProgramSources;
 
-namespace Wox.Plugin.SystemPlugins.Program
+namespace Wox.Plugins.Internal.Program
 {
     public class Program
     {
