+++ Wox.Plugin.SystemPlugins/ThirdpartyPluginIndicator.cs
@@ -3,9 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using Wox.Core;
 using Wox.Core.Data.UserSettings;
-using Wox.Plugins;
 
-namespace Wox.Plugin.SystemPlugins
+namespace Wox.Plugins.Internal
 {
     public class ThirdpartyPluginIndicator : BaseSystemPlugin
     {
