+++ Wox.Plugin.SystemPlugins/WebSearch/WebSearchSetting.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="Wox.Plugin.SystemPlugins.WebSearchSetting"
+﻿<Window x:Class="Wox.Plugins.Internal.WebSearch.WebSearchSetting"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         ResizeMode="NoResize"
