+++ Wox.Plugin.SystemPlugins/WebSearch/WebSearchesSetting.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="Wox.Plugin.SystemPlugins.WebSearchesSetting"
+﻿<UserControl x:Class="Wox.Plugins.Internal.WebSearch.WebSearchesSetting"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
