+++ Wox.Plugin.SystemPlugins/WebSearch/WebSearchesSetting.xaml.cs
@@ -1,19 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Windows;
+﻿using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 using Wox.Core.Data.UserSettings;
 
-namespace Wox.Plugin.SystemPlugins
+namespace Wox.Plugins.Internal.WebSearch
 {
     /// <summary>
     /// Interaction logic for WebSearchesSetting.xaml
@@ -46,7 +35,7 @@ namespace Wox.Plugin.SystemPlugins
 
         private void btnDeleteWebSearch_OnClick(object sender, RoutedEventArgs e)
         {
-            WebSearch seletedWebSearch = webSearchView.SelectedItem as WebSearch;
+            Core.Data.UserSettings.WebSearch seletedWebSearch = webSearchView.SelectedItem as Core.Data.UserSettings.WebSearch;
             if (seletedWebSearch != null &&
                 MessageBox.Show("Are your sure to delete " + seletedWebSearch.Title, "Delete WebSearch",
                     MessageBoxButton.YesNo) == MessageBoxResult.Yes)
@@ -62,7 +51,7 @@ namespace Wox.Plugin.SystemPlugins
 
         private void btnEditWebSearch_OnClick(object sender, RoutedEventArgs e)
         {
-            WebSearch seletedWebSearch = webSearchView.SelectedItem as WebSearch;
+            Core.Data.UserSettings.WebSearch seletedWebSearch = webSearchView.SelectedItem as Core.Data.UserSettings.WebSearch;
             if (seletedWebSearch != null)
             {
                 WebSearchSetting webSearch = new WebSearchSetting(this);
